migrate more auth tests from nose to pytest
